# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This api is the backend api responsible for the whole Coo-e coordinator app.",
    "title": "Coordinator",
    "version": "",
    "x-stoplight": {
      "id": "haridasu/specs/coordinator"
    }
  },
  "host": "resources.coo-e.com:8080",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/api/image": {
      "post": {
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "examples": {
              "application/json": {
                "image": "8a455983-cb92-4487-8641-0be97cdc8738",
                "message": "Successful",
                "status": 200
              }
            },
            "headers": {}
          },
          "400": {
            "description": "Bad Request",
            "examples": {
              "application/json": {
                "message": "Missing image data or image type.",
                "status": 400
              }
            },
            "headers": {}
          },
          "401": {
            "description": "Unauthorized",
            "examples": {
              "application/json": {
                "message": "Must be logged in.",
                "status": 401
              }
            },
            "headers": {}
          }
        },
        "tags": [],
        "description": "Lets the user upload a profile pic",
        "summary": "Upload profile pic by data"
      }
    },
    "/api/image/file/list": {
      "get": {
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "examples": {
              "application/json": {
                "files": [
                  "ZTFuTQvjgFfanRV4.jpeg"
                ],
                "message": "Success!",
                "status": 200
              }
            },
            "headers": {}
          },
          "401": {
            "description": "Unauthorized",
            "examples": {
              "application/json": {
                "message": "Must be logged in.",
                "status": 401
              }
            },
            "headers": {}
          }
        },
        "tags": [],
        "description": "shows a list of the user's images",
        "summary": "Get a list of images"
      }
    },
    "/api/image/file/upload": {
      "post": {
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$schema": "http://json-schema.org/draft-04/schema#",
              "properties": {
                "Part": {
                  "properties": {
                    "filename": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    }
                  },
                  "type": "object"
                }
              },
              "type": "object"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "examples": {
              "application/json": {
                "imageName": "{new image name with the extension}",
                "message": "Success!",
                "status": 200
              }
            },
            "headers": {}
          },
          "400": {
            "description": "Bad Request",
            "examples": {
              "application/json": {
                "message": "Please upload a png or jpeg file",
                "status": 400
              }
            },
            "headers": {}
          },
          "401": {
            "description": "Unauthorized",
            "examples": {
              "application/json": {
                "message": "Must be logged in.",
                "status": 401
              }
            },
            "headers": {}
          }
        },
        "tags": [],
        "description": "Lets the user upload an image file using multipart/form-data image upload\nthe file key or name should be \"image\". The image must be a png or jpeg",
        "summary": "Upload image file"
      }
    },
...
    "/api/image/{imageId}": {
      "get": {
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "examples": {
              "application/json": {
                "image": {
                  "createdAt": "Tue Jan 31 2017 06:16:22 GMT+0000 (UTC)",
                  "createdBy": "4542bd4b-88fe-413f-8d37-cf4b60b8189a",
                  "id": "8a455983-cb92-4487-8641-0be97cdc8738",
                  "imageData": "REALLY LONG BASE64 DATA",
                  "type": "jpeg"
                },
                "message": "Successfully retrieved from cache",
                "status": 200
              }
            },
            "headers": {}
          },
          "400": {
            "description": "Bad Request",
            "examples": {
              "application/json": {
                "message": "Unauthorized.",
                "status": 400
              }
            },
            "headers": {}
          },
          "401": {
            "description": "Unauthorized",
            "examples": {
              "application/json": {
                "message": "Must be logged in.",
                "status": 401
              }
            },
            "headers": {}
          }
        },
        "tags": [],
        "description": "",
        "summary": "Get image data by id"
      }
    },
...
  },
...
}